Move subscription page to folder
